Implemented functional database for the website
Using the database schema I implemented a database for the website to use, and called it "shopDB". It's available as the file shopDB.sql, and should be able to be queried without any expected issues. I also included separate screenshots of the database, an equivalent pdf, and a few notes on how it can be edited. It's populated with a few records for each table, and should be able to demonstrate its functionality with the website.

In addition, the database schema and subsequently the database itself was slightly  modified from before - instead of a many-to-many relationship from the Products table to the Categories table, I changed it to a many-to-one relationship from the Products table to the Categories table (so that many Products may belong to one Category). As a result, the junction table productCategories between the Products table and the Categories table was removed.