Completed shop page's front-end and functionality (shop.php)
The shop page now displays a product listing that can be filtered through user-specified search results, where search categories such as "Category", "Price", and "Rating" can influence the results returned (from the database). The product information is displayed through a set of cards numbering 3 per row (one card for each product), and can be resized depending on the browser and display size to suit a more responsive design, as well as ensuring a mobile-friendly interface. On a medium-sized display, the shop page's front-end (with the Bootstrap CSS framework) will display a set of cards numbering 2 per row, and on a smaller-sized display, the shop page's front-end will display a set of cards numbering 1 per row.

The home page's interface was also slightly modified, where the category cards in the "Shop by Category" row can now be highlighted in response to hovers from the user's mouse.

In addition, a new product page is started, named ProductPage.php, in order to provide an interface to the user for an in-depth look at a selected product and its full corresponding information (including information such as shipping speed, a return policy, and a descriptive tag offering related product information for the user).